update: added all DTOs

+++ src/main/java/coursemanagement/coursemanagement/repository/TopicRepository.java
@@ -6,5 +6,5 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface TopicRepository extends JpaRepository<Topics, Integer> {
+public interface TopicRepository extends JpaRepository<Topics, Long> {
 }
